requirements.txt
python-socketio
eventlet
gunicorn
python-dotenv

